adding refresh button to gamepage, adding survival mode to single, multiplayer, changing background image
