+++ src/components/routing/routers/AppRouter.js
@@ -1,85 +1,107 @@
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
-import { LoginGuard } from "components/routing/routeProtectors/LoginGuard";
 import { HomeGuard } from "components/routing/routeProtectors/HomeGuard";
 import HomeRouter from "components/routing/routers/HomeRouter";
 import { LobbyGuard } from "components/routing/routeProtectors/LobbyGuard";
+import { LoginGuard } from "components/routing/routeProtectors/LoginGuard";
 
-import Login from "components/views/Login";
-import Register from "components/views/Register";
-import Lobby from "components/views/home/Lobby";
-import JoinGame from "components/views/game/JoinGame";
+import Login from "components/views/authentication/Login";
+import Register from "components/views/authentication/Register";
 
-import SingleGamePreparePage from "components/views/game/SinglePlayerGame/SingleGamePreparePage";
-import SingleGamePage from "components/views/game/SinglePlayerGame/SingleGamePage";
-import SingleGameFinishPage from "components/views/game/SinglePlayerGame/GameFinishPage";
+import Lobby from "components/views/game/Lobby";
+
+import SinglePlayerGamePage from "components/views/game/SinglePlayerGame/SinglePlayerGamePage";
+import SinglePlayerGamePreparePage from "components/views/game/SinglePlayerGame/SinglePlayerGamePreparePage";
 
 import CreatedGamePage from "components/views/game/MultiPlayerGame/CreatedGamePage";
-import MutliPlayerGamePreparePage from "components/views/game/MultiPlayerGame/MultiPlayerGamePreparePage";
+import JoinGamePage from "components/views/game/MultiPlayerGame/JoinGame";
+
 import MultiPlayerGamePage from "components/views/game/MultiPlayerGame/MultiPlayerGamePage";
+import MultiPlayerGamePreparePage from "components/views/game/MultiPlayerGame/MultiPlayerGamePreparePage";
 import MultiPlayerGameFinishPage from "components/views/game/MultiPlayerGame/MultiPlayerGameFinishPage";
 
-/**
- * Main router of your application.
- * In the following class, different routes are rendered. In our case, there is a Login Route with matches the path "/login" and another Router that matches the route "/home".
- *
- * /login renders another component without any sub-route
- * /home renders a Router that contains other sub-routes that render in turn other react components
- * Documentation about routing in React: https://reacttraining.com/react-router/web/guides/quick-start
- */
+import GameFinishPage from "components/views/game/GameFinishPage";
+
+import SinglePlayerGameLobby from "components/views/game/SinglePlayerGame/SinglePlayerGameLobby"
+import MultiPlayerGameLobby from "components/views/game/MultiPlayerGame/MultiPlayerGameLobby"
+import HistoryPage from '../../views/userinfo/History';
+
+
+
 const AppRouter = () => {
-  return (
-    <BrowserRouter>
-      <Switch>
-        {/* default path */}
-        {/* <Route exact path="">
-          <Redirect to="/login" />
-        </Route> */}
-        <Route exact path="/">
-          <Redirect to="/login" />
-        </Route>
+    return (
+        <BrowserRouter>
+            <Switch>
+                {/* default path */}
+                <Route exact path="/">
+                    <Redirect to="/login" />
+                </Route>
 
-        <Route exact path="/login">
-          <LoginGuard><Login /></LoginGuard>
-        </Route>
-        <Route exact path="/register">
-          <Register />
-        </Route>
-        <Route path="/home">
-          <HomeGuard><HomeRouter base="/home" /></HomeGuard>
-        </Route>
-        <Route path="/lobby">
-          <LobbyGuard><Lobby /></LobbyGuard>
-        </Route>
+                <Route exact path="/login">
+                    <LoginGuard>
+                        <Login />
+                    </LoginGuard>
+                </Route>
+                <Route exact path={`/userinfo/history`}>
+        <HistoryPage/>
+      </Route>
+                <Route exact path="/register">
+                    <Register />
+                </Route>
 
-        <Route exact path={`/SingleGamePage/:gameId/RoundCountPage`}>
-          <SingleGamePreparePage />
-        </Route>
-        <Route exact path={`/SingleGamePage/:gameId`}>
-          <SingleGamePage />
-        </Route>
-        <Route exact path={`/SingleGamePage/:gameId/GameFinishPage`}>
-          <SingleGameFinishPage />
-        </Route>
+                <Route exact path={`/gamePage/:gameId`}>
+                    <SinglePlayerGamePage />
+                </Route>
 
-        <Route path="/StartGamePage">
-          <CreatedGamePage />
-        </Route>
-        <Route path="/JoinGame">
-          <JoinGame />
-        </Route>
-        <Route exact path={`/MultiGamePage/:gameId/RoundCountPage`}>
-          <MutliPlayerGamePreparePage />
-        </Route>
-        <Route exact path={`/MultiGamePage/:gameId`}>
-          <MultiPlayerGamePage />
-        </Route>
-        <Route exact path={`/MultiGamePage/:gameId/GameFinish`}>
-          <MultiPlayerGameFinishPage />
-        </Route>
+                <Route exact path={`/gamePage/:gameId/RounddownCountdown`}>
+                    <SinglePlayerGamePreparePage />
+                </Route>
 
-      </Switch>
-    </BrowserRouter>
-  );
+                <Route exact path={`/MultiPlayerGamePage/:gameId`}>
+                    <MultiPlayerGamePage />                    
+                </Route>
+
+                <Route exact path={`/MultiGamePage/:gameId/GameFinish`}>
+                    <MultiPlayerGameFinishPage />
+                </Route>
+
+                <Route exact path={`/MultiPlayerGamePage/:gameId/RoundCountPage`}>
+                    <MultiPlayerGamePreparePage />
+                </Route>
+
+                <Route path="/home">
+                    <HomeGuard>
+                        <HomeRouter base="/home" />
+                    </HomeGuard>
+                </Route>
+
+                <Route path="/GameFinish">
+                    <GameFinishPage>
+                        <HomeRouter base="/home" />
+                    </GameFinishPage>
+                </Route>
+
+               
+
+                <Route path="/StartGamePage">
+                    <CreatedGamePage />
+                </Route>
+
+                <Route path="/JoinGame">
+                    <JoinGamePage base="/JoinGame" />
+                </Route>
+
+                <Route path="/lobby/singleplayer">
+                    <SinglePlayerGameLobby base="/lobby/singleplayer" />
+                </Route>
+
+                <Route path="/lobby/multiplayer">
+                    <MultiPlayerGameLobby base="/lobby/multiplayer" />
+                </Route>
+               
+
+            </Switch>
+        </BrowserRouter>
+    );
 };
 
 /*
