+++ src/components/views/game/MultiPlayerGame/MultiPlayerGameFinishPage.js
@@ -2,86 +2,62 @@ import { useHistory } from "react-router-dom";
 import React, { useState, useEffect } from "react";
 import { Button } from "components/ui/Button";
 import { api, handleError } from "helpers/api";
+import InformationContainer from "components/ui/BaseContainer";
 import { Table, TableBody, TableCell, TableHead, TableRow,} from "@mui/material";
-
-import SockJS from "sockjs-client";
-import Stomp from "stompjs";
-import { getDomain } from "helpers/getDomain";
-import WebSocketType from "models/WebSocketType";
-
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+
 import "styles/views/game/FinalPage.scss";
 
 const MultiPlayerGameFinishPage = () => {
-  const [playerRanking, setPlayerRanking] = useState([]);
-  const playerId = localStorage.getItem("userId");
-  const gameId = localStorage.getItem("gameId");
-  const isServer = localStorage.getItem("isServer");
   const history = useHistory();
+  const [playerRanking, setPlayerRanking] = useState([]);
 
-  const saveGameHistory = async () => {
-    const response = await api.post(`/users/${playerId}/gameHistories/${gameId}`);
-    console.log("Game History: ", response.data);
-    toast.info(`Player's game history saved.`);
-    await new Promise((resolve) => setTimeout(resolve, 500));
-  }
-
-  async function fetchRanking() {
-    try{
-      // get the final ranking
-      const responseRanking = await api.get(`/games/${gameId}/ranking`);
-      setPlayerRanking(responseRanking.data);
-      console.log("Ranking: ", responseRanking.data);
-      await new Promise((resolve) => setTimeout(resolve, 1000));
+  const endGame = async () => {
+    if (localStorage.getItem("isServer") === 1){
+      // games/${localStorage.getItem("gameId")}
+      await api.delete(`games/${localStorage.getItem("gameId")}`);
     }
-    catch (error) {
-      toast.error("Something went wrong while fetching the ranking!");
-      console.log(handleError(error));
-    }
-  }
-  
-  useEffect(() => {
-    const Socket = new SockJS(getDomain() + "/socket");
-    const stompClient = Stomp.over(Socket);
-    let subscription;
-    stompClient.connect(
-      {},
-      (frame) => {
-        subscription = stompClient.subscribe(
-          `/instance/games/${gameId}`,
-          (message) => {
-            const messagBody = JSON.parse(message.body);
-            console.log("Socket receive msg: ", messagBody);
-            if (!isServer && messagBody.type===WebSocketType.GAME_END) {
-              saveGameHistory();
-            }
-          }
-        );
-      },
-      (err) => console.log(err)
-    );
-    return () => { subscription.unsubscribe();};
-  }, []);
+    history.push("/home");
+  };
 
   useEffect(() => {
-    async function saveGameInfo() {
+    // effect callbacks are synchronous to prevent race conditions. So we put the async function inside:
+    const saveGameHistory = async () => {
+      // ${localStorage.getItem("gameId")}
+      const response = await api.post(`/gameInfo/${localStorage.getItem("gameId")}`);
+      console.log(response.data);
+    };
+
+    const fetchData = async () => {
       try {
-        const responseGameInfo = await api.post(`/gameInfo/${gameId}`);
-        console.log("Game Info: ", responseGameInfo.data);
-        toast.info(`Game's information saved.`);
-        await new Promise((resolve) => setTimeout(resolve, 500));
+        if (localStorage.getItem("isServer") === 1) {
+          saveGameHistory();
+        }
+        // /users/${localStorage.getItem("userId")}/gameHistories/${localStorage.getItem("gameId")}
+        const responseGameInfo = await api.post(
+          `/users/${localStorage.getItem("userId")}/gameHistories/${localStorage.getItem("gameId")}`
+        );
+        console.log("gamehistory", responseGameInfo.data);
+        // get the final ranking
+        // /games/${localStorage.getItem("gameId")}/ranking
+        const responseRanking = await api.get(
+          `/games/${localStorage.getItem("gameId")}/ranking`
+        );
+        await new Promise((resolve) => setTimeout(resolve, 1000));
+        // Get the returned users and update the state.
+        setPlayerRanking(responseRanking.data);
+        console.log("player", responseRanking.data);
       }
       catch (error) {
-        toast.error("Something went wrong while fetching the ranking!");
+//        console.error(`An error occurs while saving the game history:\n${handleError(error)}`);
+//        console.error("Details:", error);
+//        alert("Something went wrong while saving the game history!");
+        toast.error("Something went wrong while saving the game history!");
         console.log(handleError(error));
       }
-    }
-    if (isServer) {
-      saveGameInfo();
-      saveGameHistory();
-    }
-    fetchRanking();
+    };
+    fetchData();
   }, []);
 
   const groupedPlayers = playerRanking.reduce((groups, player) => {
@@ -91,29 +67,13 @@ const MultiPlayerGameFinishPage = () => {
     return groups;
   }, {});
 
-  const endGame = async() => {
-    if (isServer){
-      await api.delete(`games/${gameId}`);
-      await new Promise((resolve) => setTimeout(resolve, 500));
-    }
-    localStorage.removeItem("gameId");
-    localStorage.removeItem("category");
-    localStorage.removeItem("totalRounds");
-    localStorage.removeItem("countdownTime");
-    localStorage.removeItem("isServer");
-    localStorage.removeItem("roundNumber");
-    localStorage.removeItem("myScore");
-    localStorage.removeItem("citynames");
-    localStorage.removeItem("PictureUrl");
-    localStorage.removeItem("CorrectOption");
-    history.push("/home");
-  };
 
   return (
     <div className="finalpage container">
       <h2 style={{ font: "40px" }}>
         -- Game Ended --
       </h2>
+
       <div className="podium">
         <div className="third-place">
           <div>3rd</div>
