+++ src/styles/index.scss
@@ -23,7 +23,7 @@ body {
   min-height: 100vh;
   font-family: $fontDefault;
   background-color: #0074D9; /* set background color to blue */
-  background-image: url('/img/skyline.svg');
+  background-image: url('/img/background.jpg');
   background-repeat: no-repeat;
   background-position: center center;
   background-size: cover;
@@ -66,7 +66,7 @@ button {
 }
 
 .bg-image {
-  background-image: url('/img/skyline.svg');
+  background-image: url('/img/background.jpg');
   background-repeat: no-repeat;
   background-position: center center;
   background-size: cover;
