+++ src/styles/views/game/About.scss
@@ -1,0 +1,34 @@
+@import 'styles/theme';
+
+/* The '&' is a shorthand for the containing element
+  for more information visit https://sass-lang.com/guide
+ */
+
+.about {
+    &.container {
+        background: #001F3F;
+        padding: 30px;
+        border-radius: $borderRadius;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        width: 30%;
+        margin: 50px auto 20px;
+    }
+
+    &.text-container {
+        background: #001F3F;
+        width: 50%;
+        border-radius: $borderRadius;
+        padding: 16px;
+        margin-top: 16px;
+    }
+    
+    &.button-container {
+        display: flex;
+        align-items: center;
+        margin-top: 24px;
+        flex-direction: column;
+        justify-content: center;
+    }
+}
