--- src/styles/views/home/Lobby.scss
@@ -1,178 +1,0 @@
-@import "styles/theme";
-
-/* The '&' is a shorthand for the containing element
-  for more information visit https://sass-lang.com/guide
- */
-.lobby {
-    &.layout {
-        padding: 1em;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        border-radius: $borderRadius;
-        column-gap: 1em;
-    }
-    &.container {
-        border-radius: $borderRadius;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin: 50px auto 0;
-        width: 30%;
-        padding: 30px;
-    }
-    &.container_left {
-        background: transparentize(rgb(65, 64, 64), 0.7);
-        padding: 2em;
-        background: #001F3F;
-        border-radius: $borderRadius;
-        display: flex;
-        flex-direction: column;
-        align-items: left;
-        box-shadow: $dropShadow;
-        row-gap: 2em;
-    }
-    &.container_right {
-        background: transparentize(rgb(65, 64, 64), 0.7);
-        padding: 2em;
-        border-radius: $borderRadius;
-        display: flex;
-        flex-direction: column;
-        align-items: left;
-        box-shadow: $dropShadow;
-    }
-    &.user-list {
-        padding-left: 0;
-    }
-    &.user-item {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-    }
-    &.category-select {
-        display: flex;
-        align-items: baseline;
-        justify-content: space-between;
-    }
-    &.label {
-        font-size: 20px;
-        color: $textColor;
-               
-    }
-    &.category {
-        margin-left: 20px;
-        width: 60%;
-        border-radius: $borderRadius;
-        background-color: #1979b8;
-		color: $textColor;
-        &.Mui-disabled {
-            color: $textColor;
-        }
-    }
-    &.player {
-        margin-left: 30px;
-        width: 60%;
-        border-radius: 10px;
-        background-color: #1979b8;
-		color: $textColor;
-    }
-    &.round {
-        margin-left: 30px;
-        width: 60%;
-        border-radius: $borderRadius;
-        background-color: #1979b8;
-		color: $textColor;
-        .MuiInputBase-input {
-			background-color: #1979b8;
-            color: $textColor;
-            height: 30px;
-            font-size: 14px;
-            font-weight: 400;
-            border: none;
-            border-radius: 0.5em;
-            &::placeholder {
-                color: #D4D4D4;
-              }
-        }
-    }
-
-    &.button-container {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-        row-gap: 2em;
-        column-gap: 1em;
-        
-    }
-    & p,
-    h2 {
-        font-size: 20px;
-        font-weight: 400;
-        color: $textColor;
-        margin: 0.5em 0;
-    }
-}
-
-.user {
-    &.container {
-        background: lighten($background, 5);
-        margin: 0.2em 0;
-        width: 25em;
-        padding: 10px;
-        flex-direction: column;
-        border-radius: $borderRadius;
-        display: flex;
-        align-items: left;
-    }
-    &.user-info {
-        font-weight: 600;
-        margin-left: 5px;
-        margin-right: auto;
-        margin-top: 2px;
-        color: $accent;
-    }
-}
-
-.player-name {
-    margin: auto;
-    margin-top: 20px;
-    background-color: #1979b8;
-    border-radius: 10px;
-    text-align: center;
-}
-
-.exit-button {
-    position: absolute;
-    background-color: #1979b8 !important;
-    top: 30px;
-    left: 30px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-    font-size: 2rem;
-    color: white;
-    text-transform: uppercase;
-    font-weight: bold;
-    padding: 1rem;
-    transition: all 0.3s;
-
-    &:hover {
-        background-color: #ccc;
-    }
-}
-.leader-board {
-    padding: 20px 40px;
-    & p{
-        text-align: center;
-        font-size: 28px;
-    }
-}
-
-.round-down-count {
-    border-radius: 40px !important;
-}
