+++ src/styles/views/userinfo/Profile.scss
@@ -1,0 +1,64 @@
+@import 'styles/theme';
+
+.profile {
+    &.container {
+        background: #001F3F;
+        padding: 30px;
+        border-radius: $borderRadius;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        width: 30%;
+        margin: 50px auto 0;
+    }
+
+    &.button-container {
+        display: flex;
+        justify-content: center;
+        margin-top: 2em;
+    }
+    & p,
+    h2 {
+        margin: 0.5em 0;
+        color:white;
+        background: #1979b8;
+        text-align: center;
+        width: 400px;
+        border-radius: 0.5em;
+    }
+}
+
+.user-profile-field {
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 16px;
+  border-radius: $borderRadius;
+  flex-direction: row;
+  justify-content: space-evenly;
+}
+
+.user-profile-field label {
+  font-weight: bold;
+  justify-content: flex-start; 
+  flex-basis: 30%;
+  padding: 8px;
+}
+
+.user-profile-field input {
+  border: none;
+  padding: 8px;
+  font-size: 16px;
+  background-color: #d2d9df;
+  border-radius: $borderRadius;
+  flex-basis: 67%;
+}
+
+.user-profile-field input:focus {
+  outline: none;
+}
+
+.user-profile-field value {
+  flex-basis: 70%;
+  padding: 8px;
+}
+
