+++ src/styles/views/userinfo/ScoreBoard.scss
@@ -1,0 +1,99 @@
+@import 'styles/theme';
+
+/* The '&' is a shorthand for the containing element
+  for more information visit https://sass-lang.com/guide
+ */
+.scoreboard {
+  &.container {
+    background: #001F3F;;
+    padding: 30px;
+    border-radius: $borderRadius;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 50px auto 0;
+  }
+  &.button-container {
+    display: flex;
+    width: 15em;
+    justify-content: center;
+    align-items: center;
+    margin-top: 1.5em;
+  }
+  &.label {
+    font-size: 20px;
+    font-weight: 400;
+    flex-direction: row;
+    color: $textColor;
+    display: flex;
+    width: 52%;
+}
+  & p,h2 {
+    font-size: 48px;
+    font-weight: 500;
+    margin: 0 0;
+  }
+  &.category select{
+    display: flex;
+  }
+}
+
+
+.table {
+  border-collapse: collapse;
+  border-spacing: 0;
+  color: white;
+  font-weight: bold;
+  width: 800px;
+  border-radius: 5px;
+  overflow: hidden;
+}
+.table tr:nth-child(odd) {
+  border-bottom: 1px solid #ccc;
+  background-color: #5883a795;
+}
+.table tr:nth-child(even) {
+  border-bottom: 1px solid #ccc;
+  background-color: #9ab4ca95;
+}
+.table tbody tr:hover {
+  cursor: pointer;
+  transform: translateY(-10px);
+}
+.table th {
+  padding: 8px;
+  background-color: #01010195;
+}
+.table td {
+  padding: 8px;
+  border: none;
+  text-align: center;
+}
+
+
+.user-button {
+  &:hover {
+    transform: translateY(-2px);
+  }
+  &:disabled {
+    cursor: default;
+    opacity: 0.4;
+    &:hover {
+      transform: none;
+    }
+  }
+  padding: 6px;
+  font-weight: 700;
+  text-transform: capitalize;
+  font-size: 15px;
+  text-align: center;
+  letter-spacing: 1px;
+  height: 35px;
+  border: none;
+  border-radius: $borderRadius;
+  cursor: pointer;
+  opacity: 1;
+  color: white;
+  background-color: transparent;
+  transition: all 0.3s ease;
+}
